Moved ide2csv 'info' mode to separate utility; improved info display
